Replace per-boot script with persistent network config
This setup makes some assumptions about the networking setup on the
multipass VMs – notably that there is only one network adapter and it is
named "enp0s1". With Multipass running on a MacOS host and "Jammy" now
hard-coded, I expect this will be stable for the time being (for me,
anyway).